+++ packages/gatsby-plugin-mdx/gatsby/on-create-node.js
@@ -1,18 +1,14 @@
 const fs = require(`fs`)
 const path = require(`path`)
-const crypto = require(`crypto`)
 const babel = require(`@babel/core`)
+const { createContentDigest } = require(`gatsby-core-utils`)
 
 const defaultOptions = require(`../utils/default-options`)
 const createMDXNode = require(`../utils/create-mdx-node`)
 const { MDX_SCOPES_LOCATION } = require(`../constants`)
 const genMDX = require(`../utils/gen-mdx`)
 
-const contentDigest = val =>
-  crypto
-    .createHash(`md5`)
-    .update(val)
-    .digest(`hex`)
+const contentDigest = val => createContentDigest(val)
 
 module.exports = async (
   {
@@ -125,7 +121,7 @@ class BabelPluginTransformRelativeImports {
       return {
         visitor: {
           StringLiteral({ node }) {
-            let split = node.value.split(`!`)
+            const split = node.value.split(`!`)
             const nodePath = split.pop()
             const loaders = `${split.join(`!`)}${split.length > 0 ? `!` : ``}`
             if (nodePath.startsWith(`.`)) {