+++ packages/gatsby-plugin-sharp/src/plugin-options.js
@@ -1,6 +1,6 @@
 const _ = require(`lodash`)
 
-/// Plugin options are loaded onPreBootstrap in gatsby-node
+// / Plugin options are loaded onPreBootstrap in gatsby-node
 const pluginDefaults = {
   forceBase64Format: false,
   useMozJpeg: process.env.GATSBY_JPEG_ENCODER === `MOZJPEG`,
@@ -37,13 +37,43 @@ exports.setPluginOptions = opts => {
 
 exports.getPluginOptions = () => pluginOptions
 
-const healOptions = (
+/**
+ * Creates a transform object
+ *
+ * @param {Partial<import('./process-file').TransformArgs>} args
+ */
+exports.createTransformObject = args => {
+  const options = {
+    height: args.height,
+    width: args.width,
+    cropFocus: args.cropFocus,
+    toFormat: args.toFormat,
+    pngCompressionLevel:
+      args.pngCompressionLevel || generalArgs.pngCompressionLevel,
+    quality: args.quality,
+    jpegQuality: args.jpegQuality,
+    pngQuality: args.pngQuality,
+    webpQuality: args.webpQuality,
+    jpegProgressive: args.jpegProgressive || generalArgs.jpegProgressive,
+    grayscale: args.grayscale || generalArgs.grayscale,
+    rotate: !!args.rotate,
+    trim: !!args.trim,
+    duotone: args.duotone ? args.duotone : null,
+    fit: args.fit,
+    background: args.background,
+  }
+
+  // get all non falsey values
+  return _.pickBy(options, _.identity)
+}
+
+exports.healOptions = (
   { defaultQuality: quality },
   args,
-  fileExtension,
+  fileExtension = ``,
   defaultArgs = {}
 ) => {
-  let options = _.defaults({}, args, { quality }, defaultArgs, generalArgs)
+  const options = _.defaults({}, args, { quality }, defaultArgs, generalArgs)
   options.quality = parseInt(options.quality, 10)
   options.pngCompressionLevel = parseInt(options.pngCompressionLevel, 10)
   options.pngCompressionSpeed = parseInt(options.pngCompressionSpeed, 10)
@@ -52,6 +82,11 @@ const healOptions = (
 
   // when toFormat is not set we set it based on fileExtension
   if (options.toFormat === ``) {
+    if (!fileExtension) {
+      throw new Error(
+        `toFormat seems to be empty, we need a fileExtension to set it.`
+      )
+    }
     options.toFormat = fileExtension.toLowerCase()
 
     if (fileExtension === `jpeg`) {
@@ -77,7 +112,50 @@ const healOptions = (
     options.maxHeight = parseInt(options.maxHeight, 10)
   }
 
+  ;[`width`, `height`, `maxWidth`, `maxHeight`].forEach(prop => {
+    if (typeof options[prop] !== `undefined` && options[prop] < 1) {
+      throw new Error(
+        `${prop} has to be a positive int larger than zero (> 0), now it's ${options[prop]}`
+      )
+    }
+  })
+
   return options
 }
 
-exports.healOptions = healOptions
+/**
+ * Removes all default values so we have the smallest transform args
+ *
+ * @param {Partial<import('./process-file').TransformArgs>} args
+ * @param {{defaultQuality: number }} pluginOptions
+ */
+exports.removeDefaultValues = (args, pluginOptions) => {
+  const options = {
+    height: args.height,
+    width: args.width,
+    cropFocus: args.cropFocus,
+    toFormat: args.toFormat,
+    pngCompressionLevel:
+      args.pngCompressionLevel !== generalArgs.pngCompressionLevel
+        ? args.pngCompressionLevel
+        : undefined,
+    quality:
+      args.quality !== pluginOptions.defaultQuality ? args.quality : undefined,
+    jpegQuality: args.jpegQuality,
+    pngQuality: args.pngQuality,
+    webpQuality: args.webpQuality,
+    jpegProgressive:
+      args.jpegProgressive !== generalArgs.jpegProgressive
+        ? args.jpegProgressive
+        : undefined,
+    grayscale:
+      args.grayscale !== generalArgs.grayscale ? args.grayscale : undefined,
+    rotate: args.rotate ? args.rotate : undefined,
+    trim: args.trim ? args.trim : undefined,
+    duotone: args.duotone || undefined,
+    fit: args.fit,
+    background: args.background,
+  }
+
+  return _.omitBy(options, _.isNil)
+}