+++ packages/gatsby-remark-images/src/index.js
@@ -49,7 +49,7 @@ module.exports = (
 
   // This will allow the use of html image tags
   // const rawHtmlNodes = select(markdownAST, `html`)
-  let rawHtmlNodes = []
+  const rawHtmlNodes = []
   visitWithParents(markdownAST, [`html`, `jsx`], (node, ancestors) => {
     const inLink = ancestors.some(findParentLinks)
 
@@ -57,7 +57,7 @@ module.exports = (
   })
 
   // This will only work for markdown syntax image tags
-  let markdownImageNodes = []
+  const markdownImageNodes = []
 
   visitWithParents(
     markdownAST,
@@ -150,7 +150,7 @@ module.exports = (
       return resolve()
     }
 
-    let fluidResult = await fluid({
+    const fluidResult = await fluid({
       file: imageNode,
       args: options,
       reporter,
@@ -249,6 +249,7 @@ module.exports = (
           alt="${alt}"
           title="${title}"
           loading="${loading}"
+          style="${imageStyle}"
         />
       </picture>
       `.trim()
@@ -363,7 +364,7 @@ module.exports = (
             !node.hasOwnProperty(`url`) &&
             node.hasOwnProperty(`identifier`)
           ) {
-            //consider as imageReference node
+            // consider as imageReference node
             refNode = node
             node = definitions(refNode.identifier)
             // pass original alt from referencing node
@@ -422,14 +423,14 @@ module.exports = (
               return resolve()
             }
 
-            let imageRefs = []
+            const imageRefs = []
             $(`img`).each(function() {
               imageRefs.push($(this))
             })
 
-            for (let thisImg of imageRefs) {
+            for (const thisImg of imageRefs) {
               // Get the details we need.
-              let formattedImgTag = {}
+              const formattedImgTag = {}
               formattedImgTag.url = thisImg.attr(`src`)
               formattedImgTag.title = thisImg.attr(`title`)
               formattedImgTag.alt = thisImg.attr(`alt`)