+++ packages/gatsby/cache-dir/static-entry.js
@@ -11,6 +11,7 @@ const {
   uniqBy,
   flattenDeep,
   replace,
+  concat,
 } = require(`lodash`)
 
 const apiRunner = require(`./api-runner-ssr`)
@@ -250,7 +251,7 @@ export default (pagePath, callback) => {
       const fetchKey = `assetsByChunkName[${s}]`
 
       let chunks = get(stats, fetchKey)
-      let namedChunkGroups = get(stats, `namedChunkGroups`)
+      const namedChunkGroups = get(stats, `namedChunkGroups`)
 
       if (!chunks) {
         return null
@@ -269,7 +270,7 @@ export default (pagePath, callback) => {
 
       const childAssets = namedChunkGroups[s].childAssets
       for (const rel in childAssets) {
-        chunks = merge(
+        chunks = concat(
           chunks,
           childAssets[rel].map(chunk => {
             return { rel, name: chunk }