+++ packages/gatsby/index.d.ts
@@ -703,31 +703,32 @@ export interface CreateSchemaCustomizationArgs extends ParentSpanPluginArgs {
 }
 
 export interface PreRenderHTMLArgs extends NodePluginArgs {
-  getHeadComponents: any[]
-  replaceHeadComponents: Function
-  getPreBodyComponents: any[]
-  replacePreBodyComponents: Function
-  getPostBodyComponents: any[]
-  replacePostBodyComponents: Function
+  getHeadComponents: () => React.ReactNode[]
+  replaceHeadComponents: (comp: React.ReactNode[]) => void
+  getPreBodyComponents: () => React.ReactNode[]
+  replacePreBodyComponents: (comp: React.ReactNode[]) => void
+  getPostBodyComponents: () => React.ReactNode[]
+  replacePostBodyComponents: (comp: React.ReactNode[]) => void
 }
 
+type ReactProps<T extends Element> = React.DetailedHTMLProps<React.HTMLAttributes<T>, T>
 export interface RenderBodyArgs extends NodePluginArgs {
   pathname: string
-  setHeadComponents: Function
-  setHtmlAttributes: Function
-  setBodyAttributes: Function
-  setPreBodyComponents: Function
-  setPostBodyComponents: Function
+  setHeadComponents: (comp: React.ReactNode[]) => void
+  setHtmlAttributes: (attr: ReactProps<HTMLHtmlElement>) => void
+  setBodyAttributes: (attr: ReactProps<HTMLBodyElement>) => void
+  setPreBodyComponents: (comp: React.ReactNode[]) => void
+  setPostBodyComponents: (comp: React.ReactNode[]) => void
   setBodyProps: Function
 }
 
 export interface ReplaceRendererArgs extends NodePluginArgs {
-  replaceBodyHTMLString: Function
-  setHeadComponents: Function
-  setHtmlAttributes: Function
-  setBodyAttributes: Function
-  setPreBodyComponents: Function
-  setPostBodyComponents: Function
+  replaceBodyHTMLString: (str: string) => void
+  setHeadComponents: (comp: React.ReactNode[]) => void
+  setHtmlAttributes: (attr: ReactProps<HTMLHtmlElement>) => void
+  setBodyAttributes: (attr: ReactProps<HTMLBodyElement>) => void
+  setPreBodyComponents: (comp: React.ReactNode[]) => void
+  setPostBodyComponents: (comp: React.ReactNode[]) => void
   setBodyProps: Function
 }
 
@@ -803,7 +804,12 @@ export interface Actions {
 
   /** @see https://www.gatsbyjs.org/docs/actions/#createPage */
   createPage<TContext = Record<string, unknown>>(
-    args: { path: string; matchPath?: string; component: string; context: TContext },
+    args: {
+      path: string
+      matchPath?: string
+      component: string
+      context: TContext
+    },
     plugin?: ActionPlugin,
     option?: ActionOptions
   ): void
@@ -876,6 +882,17 @@ export interface Actions {
     job: Record<string, unknown> & { id: string },
     plugin?: ActionPlugin
   ): void
+
+  /** @see https://www.gatsbyjs.org/docs/actions/#createJobV2 */
+  createJobV2(
+    job: {
+      name: string
+      inputPaths: string[]
+      outputDir: string
+      args: Record<string, unknown>
+    },
+    plugin?: ActionPlugin
+  ): Promise<unknown>
 
   /** @see https://www.gatsbyjs.org/docs/actions/#setJob */
   setJob(