+++ packages/gatsby-page-utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.37](https://github.com/gatsbyjs/gatsby/compare/gatsby-page-utils@0.0.36...gatsby-page-utils@0.0.37) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-page-utils
+
+## [0.0.36](https://github.com/gatsbyjs/gatsby/compare/gatsby-page-utils@0.0.35...gatsby-page-utils@0.0.36) (2019-12-20)
+
+**Note:** Version bump only for package gatsby-page-utils
+
 ## [0.0.35](https://github.com/gatsbyjs/gatsby/compare/gatsby-page-utils@0.0.33...gatsby-page-utils@0.0.35) (2019-12-10)
 
 **Note:** Version bump only for package gatsby-page-utils