+++ packages/gatsby-plugin-sharp/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.3.13...gatsby-plugin-sharp@2.4.0) (2020-01-21)
+
+### Features
+
+- **gatsby-plugin-sharp:** move to jobs-api v2 ([#19940](https://github.com/gatsbyjs/gatsby/issues/19940)) ([dbab3f9](https://github.com/gatsbyjs/gatsby/commit/dbab3f9))
+
+## [2.3.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.3.12...gatsby-plugin-sharp@2.3.13) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.3.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.3.10...gatsby-plugin-sharp@2.3.12) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.3.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.3.10...gatsby-plugin-sharp@2.3.11) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.3.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.3.9...gatsby-plugin-sharp@2.3.10) (2019-12-20)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.3.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.3.8...gatsby-plugin-sharp@2.3.9) (2019-12-17)
+
+### Bug Fixes
+
+- **#19220:** added default args for generateBase64 function ([#20158](https://github.com/gatsbyjs/gatsby/issues/20158)) ([d6c2b73](https://github.com/gatsbyjs/gatsby/commit/d6c2b73)), closes [#19220](https://github.com/gatsbyjs/gatsby/issues/19220)
+
+## [2.3.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.3.7...gatsby-plugin-sharp@2.3.8) (2019-12-16)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
 ## [2.3.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.3.5...gatsby-plugin-sharp@2.3.7) (2019-12-10)
 
 **Note:** Version bump only for package gatsby-plugin-sharp