+++ packages/gatsby-plugin-stylus/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-stylus@2.1.19...gatsby-plugin-stylus@2.1.20) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-plugin-stylus
+
 ## [2.1.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-stylus@2.1.17...gatsby-plugin-stylus@2.1.19) (2019-12-10)
 
 **Note:** Version bump only for package gatsby-plugin-stylus