+++ packages/gatsby-remark-embed-snippet/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-embed-snippet@4.1.20...gatsby-remark-embed-snippet@4.1.21) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-remark-embed-snippet
+
+## [4.1.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-embed-snippet@4.1.18...gatsby-remark-embed-snippet@4.1.20) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-remark-embed-snippet
+
+## [4.1.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-embed-snippet@4.1.18...gatsby-remark-embed-snippet@4.1.19) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-remark-embed-snippet
+
 ## [4.1.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-embed-snippet@4.1.16...gatsby-remark-embed-snippet@4.1.18) (2019-12-10)
 
 **Note:** Version bump only for package gatsby-remark-embed-snippet